+++ packages/core/src/common/util/migration/migrations/migrateToV5.ts
@@ -1,27 +1,4 @@
-import {
-  CategoricalVariableType,
-  DataPointType,
-  ExperimentResultType,
-  OptimizerConfig,
-  ScoreVariableType,
-  ValueVariableType,
-  Info,
-} from '@core/common/types'
-
-type ExperimentTypeV5 = {
-  id: string
-  changedSinceLastEvaluation: boolean
-  info: Info
-  extras: Record<string, unknown>
-  categoricalVariables: CategoricalVariableType[]
-  valueVariables: ValueVariableType[]
-  scoreVariables: ScoreVariableType[]
-  optimizerConfig: OptimizerConfig
-  results: ExperimentResultType
-  dataPoints: DataPointType[][]
-}
-
-export const migrateToV5 = (json: ExperimentTypeV5): ExperimentTypeV5 => {
+export const migrateToV5 = (json: any): any => {
   return {
     ...json,
     scoreVariables: [
@@ -31,8 +8,8 @@ export const migrateToV5 = (json: ExperimentTypeV5): ExperimentTypeV5 => {
         enabled: true,
       },
     ],
-    dataPoints: json.dataPoints.map(dps =>
-      dps.map(dp =>
+    dataPoints: json.dataPoints.map((dps: any) =>
+      dps.map((dp: any) =>
         dp.name === 'score' && Array.isArray(dp.value)
           ? { ...dp, value: dp.value[0] ?? 0 }
           : dp