+++ packages/core/src/context/experiment/experiment-context.test.tsx
@@ -1,13 +1,15 @@
 import { ApiProvider } from '@core/context/experiment/api-provider'
-import { renderHook } from '@testing-library/react'
+import { render, renderHook, screen } from '@testing-library/react'
 import { FC } from 'react'
-import { vi } from 'vitest'
+import { it, vi } from 'vitest'
 import {
   useExperiment,
   ExperimentProvider,
   useSelector,
 } from './experiment-context'
 import { State } from './store'
+
+import catapult from '@core/sample-data/catapult.json'
 
 const ExperimentWrapper: FC<{ children: React.ReactNode }> = ({ children }) => (
   <ApiProvider>
@@ -39,5 +41,65 @@ describe('useSelector', () => {
       wrapper: ExperimentWrapper,
     })
     expect(result.current).toEqual('123')
+  })
+})
+
+describe('Pluggable storage', () => {
+  afterEach(() => localStorage.clear())
+  const Tester = () => {
+    const { state, loading } = useExperiment()
+    return (
+      <pre>
+        {loading ? 'loading' : 'done'}
+        {JSON.stringify(state.experiment, undefined, 2)}
+      </pre>
+    )
+  }
+
+  it('should read from localStorage when no storage provider is assigned', async () => {
+    const localCatapult: State = {
+      experiment: JSON.parse(JSON.stringify(catapult)),
+    }
+    localCatapult.experiment.info.name = 'I am stored in localStorage'
+    localCatapult.experiment.id = 'catapult'
+    localStorage.setItem('catapult', JSON.stringify(localCatapult))
+    render(
+      <ApiProvider>
+        <ExperimentProvider experimentId="catapult">
+          <Tester />
+        </ExperimentProvider>
+      </ApiProvider>
+    )
+    const experimentId = await screen.findAllByText(/catapult/)
+    const experimentName = await screen.findAllByText(
+      /I am stored in localStorage/
+    )
+    expect(experimentId).not.toBeNull()
+    expect(experimentName).not.toBeNull()
+  })
+
+  it('should read from provided storage when such provider is assigned', async () => {
+    const localCatapult: State = {
+      experiment: JSON.parse(JSON.stringify(catapult)),
+    }
+    localCatapult.experiment.info.name = 'I am stored in storage provider'
+    localCatapult.experiment.id = 'catapult'
+    const storage: Storage = {
+      ...localStorage,
+      getItem: () => JSON.stringify(localCatapult),
+    }
+    render(
+      <ApiProvider>
+        <ExperimentProvider storage={storage} experimentId="catapult">
+          <Tester />
+        </ExperimentProvider>
+      </ApiProvider>
+    )
+    const experimentId = await screen.findAllByText(/catapult/)
+    const experimentName = await screen.findAllByText(
+      /I am stored in storage provider/
+    )
+    expect(experimentId).not.toBeNull()
+    expect(experimentName).not.toBeNull()
   })
 })