+++ packages/core/src/context/experiment/experiment-context.tsx
@@ -1,15 +1,11 @@
 import * as React from 'react'
 import { useLocalStorageReducer } from '@core/storage'
-import {
-  DefaultApi,
-  OptimizerapiOptimizerRunRequest,
-} from '@boostv/process-optimizer-frontend-api'
+import { DefaultApi } from '@boostv/process-optimizer-frontend-api'
 import { Dispatch, rootReducer } from './reducers'
-import { calculateData, calculateSpace } from '@core/common/'
 import { migrate } from '@core/common'
 import { initialState, State, useApi } from '@core/context/experiment'
-import { ExperimentResultType, ExperimentType } from '@core/common/types'
-import { versionInfo } from '@core/common'
+import { ExperimentType } from '@core/common/types'
+import { fetchExperimentResult } from '@core/context/experiment/api'
 
 const ExperimentContext = React.createContext<
   | {
@@ -23,14 +19,15 @@ const ExperimentContext = React.createContext<
 
 type ExperimentProviderProps = {
   experimentId: string
-  children: any
+  children?: React.ReactNode
+  storage?: Storage
 }
 
-export function ExperimentProvider({
+export const ExperimentProvider = ({
   experimentId,
   children,
-}: ExperimentProviderProps) {
-  const api = useApi()
+  storage,
+}: ExperimentProviderProps) => {
   const storageKey = experimentId === undefined ? 'unknown' : experimentId
   const initialExperimentState = {
     ...initialState,
@@ -40,16 +37,30 @@ export function ExperimentProvider({
     rootReducer,
     initialExperimentState,
     storageKey,
-    (a: State) => ({ ...a, experiment: migrate(a.experiment) })
+    (a: State) => ({ ...a, experiment: migrate(a.experiment) }),
+    storage
   )
-  const [loading, setLoading] = React.useState(true)
+  return (
+    <ManagedExperimentProvider state={state} dispatch={dispatch}>
+      {children}
+    </ManagedExperimentProvider>
+  )
+}
 
-  React.useEffect(() => {
-    if (state?.experiment?.info?.swVersion !== versionInfo.version) {
-      dispatch({ type: 'setSwVersion', payload: versionInfo.version })
-    }
-    setLoading(false)
-  }, [dispatch, experimentId, state])
+export type ManagedExperimentProviderProps = {
+  state: State
+  dispatch: Dispatch
+  children: React.ReactNode
+}
+
+export function ManagedExperimentProvider({
+  state,
+  dispatch,
+  children,
+}: ManagedExperimentProviderProps) {
+  const api = useApi()
+
+  const [loading, setLoading] = React.useState(false)
 
   const getValue = (callback: (state: State) => any) => callback(state)
 
@@ -58,8 +69,18 @@ export function ExperimentProvider({
     dispatch,
     getValue,
     loading,
-    evaluate: (exp?: ExperimentType) =>
-      runExperiment(dispatch, exp ?? state.experiment, api),
+    evaluate: async (exp?: ExperimentType) => {
+      if (!loading) {
+        setLoading(true)
+        try {
+          await runExperiment(dispatch, exp ?? state.experiment, api)
+        } catch (e) {
+          setLoading(false)
+          throw e
+        }
+        setLoading(false)
+      }
+    },
   }
   return (
     <ExperimentContext.Provider value={value}>
@@ -82,51 +103,6 @@ export const useSelector = <T,>(selector: (state: State) => T) => {
     throw new Error('useSelector must be used within an ExperimentProvider')
   }
   return selector(context.state)
-}
-
-const fetchExperimentResult = async (
-  experiment: ExperimentType,
-  api: DefaultApi
-): Promise<ExperimentResultType> => {
-  const cfg = experiment.optimizerConfig
-  const extras = experiment.extras || {}
-  const space = calculateSpace(experiment)
-
-  const request: OptimizerapiOptimizerRunRequest = {
-    experiment: {
-      data: calculateData(
-        experiment.categoricalVariables,
-        experiment.valueVariables,
-        experiment.scoreVariables,
-        experiment.dataPoints
-      ),
-      extras: extras,
-      optimizerConfig: {
-        acqFunc: cfg.acqFunc,
-        baseEstimator: cfg.baseEstimator,
-        initialPoints: Number(cfg.initialPoints),
-        kappa: Number(cfg.kappa),
-        xi: Number(cfg.xi),
-        space: space,
-      },
-    },
-  }
-
-  const result = await api.optimizerapiOptimizerRun(request)
-
-  const experimentResult: ExperimentResultType = {
-    id: experiment.id,
-    plots:
-      result.plots?.map(p => ({ id: p.id ?? '', plot: p.plot ?? '' })) ?? [],
-    // TODO: Remove cast, use zod
-    next: (result.result?.next ?? []) as (string | number)[][],
-    pickled: result.result?.pickled ?? '',
-    expectedMinimum:
-      result.result?.models?.find(() => true)?.expectedMinimum ?? [],
-    extras: result.result?.extras ?? {},
-  }
-
-  return experimentResult
 }
 
 async function runExperiment(