+++ packages/core/src/storage/useLocalStorageReducer.ts
@@ -1,16 +1,15 @@
 import { useReducer } from 'react'
 
 const init =
-  (localStorageKey: string) =>
+  (localStorageKey: string, storage: typeof localStorage = localStorage) =>
   <S>(initialState: S) => {
     try {
-      const locallyStoredState = localStorage.getItem(localStorageKey)
+      const locallyStoredState = storage.getItem(localStorageKey)
       if (locallyStoredState !== null) {
-        console.log(`Found data in ${localStorageKey}`)
         const state = JSON.parse(locallyStoredState)
         return { ...initialState, ...state }
       } else {
-        localStorage.setItem(localStorageKey, JSON.stringify(initialState))
+        storage.setItem(localStorageKey, JSON.stringify(initialState))
       }
     } catch (error) {
       // Incognito mode might cause loading to fail - add error message to initial state and continue
@@ -22,18 +21,19 @@ export const useLocalStorageReducer = <S, A>(
   reducer: (state: S, action: A) => S,
   initialState: S,
   localStorageKey = 'rootState',
-  transform = (x: S) => x
+  transform = (x: S) => x,
+  storage: typeof localStorage = localStorage
 ) => {
   const localStorageReducer = (state: S, action: A) => {
     const newState = reducer(state, action)
     try {
-      localStorage.setItem(localStorageKey, JSON.stringify(newState))
+      storage.setItem(localStorageKey, JSON.stringify(newState))
     } catch {
       console.log(`Unable to use local storage`)
     }
     return newState
   }
   return useReducer(localStorageReducer, initialState, s =>
-    transform(init(localStorageKey)(s))
+    transform(init(localStorageKey, storage)(s))
   )
 }