+++ packages/ui/src/containers/result-data/experimentation-guide.style.ts
@@ -1,5 +1,5 @@
 import { makeStyles } from 'tss-react/mui'
-import { grey } from '@mui/material/colors'
+import { colors } from '@mui/material'
 
 export const useStyles = makeStyles()(() => ({
   titleButton: {
@@ -9,7 +9,7 @@ export const useStyles = makeStyles()(() => ({
     color: 'white',
   },
   extrasContainer: {
-    background: grey[200],
+    background: colors.grey[200],
   },
 }))