+++ packages/ui/src/features/input-model/input-model.style.ts
@@ -1,9 +1,9 @@
 import { makeStyles } from 'tss-react/mui'
-import { grey } from '@mui/material/colors'
+import { colors } from '@mui/material'
 
 export const useStyles = makeStyles()(() => ({
   editBox: {
-    background: grey[200],
+    background: colors.grey[200],
   },
   iconValueType: {
     fontSize: 10,