+++ packages/ui/vite.config.ts
@@ -1,17 +1,24 @@
 /// <reference types="vitest" />
-import { defineConfig } from 'vite'
+import { defineConfig, PluginOption } from 'vite'
 import { resolve } from 'path'
 import react from '@vitejs/plugin-react'
 import dts from 'vite-plugin-dts'
+import replace from '@rollup/plugin-replace'
 
 export default defineConfig({
-  plugins: [dts(), react()],
+  plugins: [
+    react({
+      jsxRuntime: 'classic',
+    }),
+    dts(),
+    replace({
+      'mui/material/styles': 'mui/material',
+    }),
+  ],
   resolve: {
     alias: [{ find: '@ui', replacement: resolve(__dirname, './src') }],
   },
   build: {
-    sourcemap: true,
-    target: 'esnext',
     minify: false,
     lib: {
       entry: resolve(__dirname, 'src/index.ts'),
@@ -22,13 +29,13 @@ export default defineConfig({
         'react',
         'react-dom',
         'react-hook-form',
+        '@boostv/process-optimizer-frontend-api',
         '@boostv/process-optimizer-frontend-core',
-        '@mui/icons-material',
-        '@mui/material',
+        '@boostv/process-optimizer-frontend-plots',
+        'immer',
+        /@mui.*/,
         '@emotion/react',
         '@emotion/styled',
-        'tss-react',
-        'tss-react/mui',
       ],
     },
     outDir: 'dist',