+++ sample-app/src/context/global/global-reducer.ts
@@ -1,6 +1,6 @@
 import { ThemeName } from '@sample/theme/theme'
 import { assertUnreachable } from '@boostv/process-optimizer-frontend-core'
-import produce from 'immer'
+import { produce } from 'immer'
 
 export type State = {
   debug: boolean