+++ packages/plots/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @process-optimizer-frontend/plots
 
+## 1.0.3
+
+### Patch Changes
+
+- 28a3fd7: Fix bundling error
+- a912e3f: Convert plots to ViteJS
+
+## 1.0.2
+
+### Patch Changes
+
+- 739e6b8: Revert conversion to tsc and use vite for all packages
+
+## 1.0.1
+
+### Patch Changes
+
+- 58fbac8: Use plain tsc for building packages
+
 ## 1.0.0
 
 ### Patch Changes