+++ sample-app/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @process-optimizer-frontend/sample-app
 
+## 1.9.0
+
+### Minor Changes
+
+- f915c8a: Introduce Zod for input validation
+
+### Patch Changes
+
+- Updated dependencies [f915c8a]
+  - @boostv/process-optimizer-frontend-core@1.4.0
+  - @boostv/process-optimizer-frontend-ui@1.2.0
+
+## 1.8.3
+
+### Patch Changes
+
+- 28a3fd7: Fix bundling error
+- Updated dependencies [28a3fd7]
+- Updated dependencies [a912e3f]
+  - @boostv/process-optimizer-frontend-plots@1.0.3
+  - @boostv/process-optimizer-frontend-ui@1.1.5
+
+## 1.8.2
+
+### Patch Changes
+
+- 58fbac8: Use plain tsc for building packages
+- Updated dependencies [58fbac8]
+  - @boostv/process-optimizer-frontend-plots@1.0.1
+  - @boostv/process-optimizer-frontend-core@1.3.2
+  - @boostv/process-optimizer-frontend-ui@1.1.2
+
+## 1.8.1
+
+### Patch Changes
+
+- 123d09d: Fix various minor issues related to API provider
+- 53172b5: Minor fix to API provider
+- Updated dependencies [842cb41]
+- Updated dependencies [123d09d]
+- Updated dependencies [53172b5]
+  - @boostv/process-optimizer-frontend-core@1.2.0
+  - @boostv/process-optimizer-frontend-ui@1.1.1
+
 ## 1.8.0
 
 ### Minor Changes